# Improved performance of take by 50